기능: 설정 뷰 에서 앱 버전 뷰 분리
